+++ overlord/ifacestate/ifacestate_test.go
@@ -41,6 +41,8 @@ import (
 	"github.com/snapcore/snapd/logger"
 	"github.com/snapcore/snapd/osutil"
 	"github.com/snapcore/snapd/overlord"
+	"github.com/snapcore/snapd/overlord/assertstate"
+	"github.com/snapcore/snapd/overlord/auth"
 	"github.com/snapcore/snapd/overlord/hookstate"
 	"github.com/snapcore/snapd/overlord/ifacestate"
 	"github.com/snapcore/snapd/overlord/ifacestate/ifacerepo"
@@ -55,9 +57,119 @@ import (
 
 func TestInterfaceManager(t *testing.T) { TestingT(t) }
 
+type AssertsMock struct {
+	Db           *asserts.Database
+	storeSigning *assertstest.StoreStack
+	brandSigning *assertstest.SigningDB
+}
+
+func (am *AssertsMock) SetupAsserts(c *C, st *state.State) {
+	am.storeSigning = assertstest.NewStoreStack("canonical", nil)
+	brandPrivKey, _ := assertstest.GenerateKey(752)
+	am.brandSigning = assertstest.NewSigningDB("my-brand", brandPrivKey)
+
+	db, err := asserts.OpenDatabase(&asserts.DatabaseConfig{
+		Backstore: asserts.NewMemoryBackstore(),
+		Trusted:   am.storeSigning.Trusted,
+	})
+	c.Assert(err, IsNil)
+	am.Db = db
+	err = db.Add(am.storeSigning.StoreAccountKey(""))
+	c.Assert(err, IsNil)
+
+	st.Lock()
+	assertstate.ReplaceDB(st, am.Db)
+
+	brandAcct := assertstest.NewAccount(am.storeSigning, "my-brand", map[string]interface{}{
+		"account-id": "my-brand",
+	}, "")
+	err = assertstate.Add(st, brandAcct)
+	c.Assert(err, IsNil)
+
+	brandPubKey, err := am.brandSigning.PublicKey("")
+	c.Assert(err, IsNil)
+	brandAccKey := assertstest.NewAccountKey(am.storeSigning, brandAcct, nil, brandPubKey, "")
+	err = assertstate.Add(st, brandAccKey)
+	c.Assert(err, IsNil)
+	st.Unlock()
+}
+
+func (am *AssertsMock) MockModel(c *C, st *state.State, extraHeaders map[string]interface{}) {
+	headers := map[string]interface{}{
+		"series":       "16",
+		"brand-id":     "my-brand",
+		"model":        "my-model",
+		"gadget":       "gadget",
+		"kernel":       "krnl",
+		"architecture": "amd64",
+		"timestamp":    time.Now().Format(time.RFC3339),
+	}
+	for k, v := range extraHeaders {
+		headers[k] = v
+	}
+	model, err := am.brandSigning.Sign(asserts.ModelType, headers, nil, "")
+	c.Assert(err, IsNil)
+	st.Lock()
+	defer st.Unlock()
+	err = assertstate.Add(st, model)
+	c.Assert(err, IsNil)
+	err = auth.SetDevice(st, &auth.DeviceState{
+		Brand: "my-brand",
+		Model: "my-model",
+	})
+	c.Assert(err, IsNil)
+}
+
+func (am *AssertsMock) MockSnapDecl(c *C, name, publisher string, extraHeaders map[string]interface{}) {
+	_, err := am.Db.Find(asserts.AccountType, map[string]string{
+		"account-id": publisher,
+	})
+	if asserts.IsNotFound(err) {
+		acct := assertstest.NewAccount(am.storeSigning, publisher, map[string]interface{}{
+			"account-id": publisher,
+		}, "")
+		err = am.Db.Add(acct)
+	}
+	c.Assert(err, IsNil)
+
+	headers := map[string]interface{}{
+		"series":       "16",
+		"snap-name":    name,
+		"publisher-id": publisher,
+		"snap-id":      (name + strings.Repeat("id", 16))[:32],
+		"timestamp":    time.Now().Format(time.RFC3339),
+	}
+	for k, v := range extraHeaders {
+		headers[k] = v
+	}
+
+	snapDecl, err := am.storeSigning.Sign(asserts.SnapDeclarationType, headers, nil, "")
+	c.Assert(err, IsNil)
+
+	err = am.Db.Add(snapDecl)
+	c.Assert(err, IsNil)
+}
+
+func (am *AssertsMock) MockStore(c *C, st *state.State, storeID string, extraHeaders map[string]interface{}) {
+	headers := map[string]interface{}{
+		"store":       storeID,
+		"operator-id": am.storeSigning.AuthorityID,
+		"timestamp":   time.Now().Format(time.RFC3339),
+	}
+	for k, v := range extraHeaders {
+		headers[k] = v
+	}
+	storeAs, err := am.storeSigning.Sign(asserts.StoreType, headers, nil, "")
+	c.Assert(err, IsNil)
+	st.Lock()
+	defer st.Unlock()
+	err = assertstate.Add(st, storeAs)
+	c.Assert(err, IsNil)
+}
+
 type interfaceManagerSuite struct {
 	testutil.BaseTest
-	ifacestate.AssertsMock
+	AssertsMock
 	o              *overlord.Overlord
 	state          *state.State
 	se             *overlord.StateEngine
@@ -83,7 +195,7 @@ func (s *interfaceManagerSuite) SetUpTest(c *C) {
 	s.state = s.o.State()
 	s.se = s.o.StateEngine()
 
-	s.AssertsMock.MockAsserts(c, s.state)
+	s.SetupAsserts(c, s.state)
 
 	s.BaseTest.AddCleanup(snap.MockSanitizePlugsSlots(func(snapInfo *snap.Info) {}))
 
@@ -680,7 +792,7 @@ func (s *interfaceManagerSuite) TestAutoconnectIgnoresSetupProfilesPhase2(c *C) 
 }
 
 func (s *interfaceManagerSuite) TestEnsureProcessesConnectTask(c *C) {
-	s.AssertsMock.MockModel(c, s.state, nil)
+	s.MockModel(c, s.state, nil)
 
 	s.mockIfaces(c, &ifacetest.TestInterface{InterfaceName: "test"}, &ifacetest.TestInterface{InterfaceName: "test2"})
 	s.mockSnap(c, consumerYaml)
@@ -786,12 +898,12 @@ func (s *interfaceManagerSuite) TestConnectTaskNoSuchPlug(c *C) {
 }
 
 func (s *interfaceManagerSuite) TestConnectTaskCheckNotAllowed(c *C) {
-	s.AssertsMock.MockModel(c, s.state, nil)
+	s.MockModel(c, s.state, nil)
 
 	s.testConnectTaskCheck(c, func() {
-		s.AssertsMock.MockSnapDecl(c, "consumer", "consumer-publisher", nil)
+		s.MockSnapDecl(c, "consumer", "consumer-publisher", nil)
 		s.mockSnap(c, consumerYaml)
-		s.AssertsMock.MockSnapDecl(c, "producer", "producer-publisher", nil)
+		s.MockSnapDecl(c, "producer", "producer-publisher", nil)
 		s.mockSnap(c, producerYaml)
 	}, func(change *state.Change) {
 		c.Check(change.Err(), ErrorMatches, `(?s).*connection not allowed by slot rule of interface "test".*`)
@@ -804,7 +916,7 @@ func (s *interfaceManagerSuite) TestConnectTaskCheckNotAllowed(c *C) {
 }
 
 func (s *interfaceManagerSuite) TestConnectTaskCheckNotAllowedButNoDecl(c *C) {
-	s.AssertsMock.MockModel(c, s.state, nil)
+	s.MockModel(c, s.state, nil)
 
 	s.testConnectTaskCheck(c, func() {
 		s.mockSnap(c, consumerYaml)
@@ -821,12 +933,12 @@ func (s *interfaceManagerSuite) TestConnectTaskCheckNotAllowedButNoDecl(c *C) {
 }
 
 func (s *interfaceManagerSuite) TestConnectTaskCheckAllowed(c *C) {
-	s.AssertsMock.MockModel(c, s.state, nil)
+	s.MockModel(c, s.state, nil)
 
 	s.testConnectTaskCheck(c, func() {
-		s.AssertsMock.MockSnapDecl(c, "consumer", "one-publisher", nil)
+		s.MockSnapDecl(c, "consumer", "one-publisher", nil)
 		s.mockSnap(c, consumerYaml)
-		s.AssertsMock.MockSnapDecl(c, "producer", "one-publisher", nil)
+		s.MockSnapDecl(c, "producer", "one-publisher", nil)
 		s.mockSnap(c, producerYaml)
 	}, func(change *state.Change) {
 		c.Assert(change.Err(), IsNil)
@@ -879,7 +991,7 @@ slots:
 }
 
 func (s *interfaceManagerSuite) TestConnectTaskCheckDeviceScopeNoStore(c *C) {
-	s.AssertsMock.MockModel(c, s.state, nil)
+	s.MockModel(c, s.state, nil)
 
 	s.testConnectTaskCheckDeviceScope(c, func(change *state.Change) {
 		c.Check(change.Err(), ErrorMatches, `(?s).*connection not allowed by plug rule of interface "test".*`)
@@ -892,7 +1004,7 @@ func (s *interfaceManagerSuite) TestConnectTaskCheckDeviceScopeNoStore(c *C) {
 }
 
 func (s *interfaceManagerSuite) TestConnectTaskCheckDeviceScopeWrongStore(c *C) {
-	s.AssertsMock.MockModel(c, s.state, map[string]interface{}{
+	s.MockModel(c, s.state, map[string]interface{}{
 		"store": "other-store",
 	})
 
@@ -907,7 +1019,7 @@ func (s *interfaceManagerSuite) TestConnectTaskCheckDeviceScopeWrongStore(c *C) 
 }
 
 func (s *interfaceManagerSuite) TestConnectTaskCheckDeviceScopeRightStore(c *C) {
-	s.AssertsMock.MockModel(c, s.state, map[string]interface{}{
+	s.MockModel(c, s.state, map[string]interface{}{
 		"store": "my-store",
 	})
 
@@ -923,11 +1035,11 @@ func (s *interfaceManagerSuite) TestConnectTaskCheckDeviceScopeRightStore(c *C) 
 }
 
 func (s *interfaceManagerSuite) TestConnectTaskCheckDeviceScopeWrongFriendlyStore(c *C) {
-	s.AssertsMock.MockModel(c, s.state, map[string]interface{}{
+	s.MockModel(c, s.state, map[string]interface{}{
 		"store": "my-substore",
 	})
 
-	s.AssertsMock.MockStore(c, s.state, "my-substore", map[string]interface{}{
+	s.MockStore(c, s.state, "my-substore", map[string]interface{}{
 		"friendly-stores": []interface{}{"other-store"},
 	})
 
@@ -942,11 +1054,11 @@ func (s *interfaceManagerSuite) TestConnectTaskCheckDeviceScopeWrongFriendlyStor
 }
 
 func (s *interfaceManagerSuite) TestConnectTaskCheckDeviceScopeRightFriendlyStore(c *C) {
-	s.AssertsMock.MockModel(c, s.state, map[string]interface{}{
+	s.MockModel(c, s.state, map[string]interface{}{
 		"store": "my-substore",
 	})
 
-	s.AssertsMock.MockStore(c, s.state, "my-substore", map[string]interface{}{
+	s.MockStore(c, s.state, "my-substore", map[string]interface{}{
 		"friendly-stores": []interface{}{"my-store"},
 	})
 
@@ -973,9 +1085,9 @@ slots:
 	defer restore()
 	s.mockIfaces(c, &ifacetest.TestInterface{InterfaceName: "test"})
 
-	s.AssertsMock.MockSnapDecl(c, "producer", "one-publisher", nil)
+	s.MockSnapDecl(c, "producer", "one-publisher", nil)
 	s.mockSnap(c, producerYaml)
-	s.AssertsMock.MockSnapDecl(c, "consumer", "one-publisher", map[string]interface{}{
+	s.MockSnapDecl(c, "consumer", "one-publisher", map[string]interface{}{
 		"format": "3",
 		"plugs": map[string]interface{}{
 			"test": map[string]interface{}{
@@ -1308,7 +1420,7 @@ func (s *interfaceManagerSuite) mockSnapInstance(c *C, instanceName, yamlText st
 	snapInfo := snaptest.MockSnapInstance(c, instanceName, yamlText, sideInfo)
 	sideInfo.RealName = snapInfo.SnapName()
 
-	a, err := s.AssertsMock.Db.FindMany(asserts.SnapDeclarationType, map[string]string{
+	a, err := s.Db.FindMany(asserts.SnapDeclarationType, map[string]string{
 		"snap-name": sideInfo.RealName,
 	})
 	if err == nil {
@@ -1537,7 +1649,7 @@ func (s *interfaceManagerSuite) TestDoSetupSnapSecurityHonorsUndesiredFlag(c *C)
 
 // The auto-connect task will auto-connect plugs with viable candidates.
 func (s *interfaceManagerSuite) TestDoSetupSnapSecurityAutoConnectsPlugs(c *C) {
-	s.AssertsMock.MockModel(c, s.state, nil)
+	s.MockModel(c, s.state, nil)
 
 	// Add an OS snap.
 	s.mockSnap(c, ubuntuCoreSnapYaml)
@@ -1583,7 +1695,7 @@ func (s *interfaceManagerSuite) TestDoSetupSnapSecurityAutoConnectsPlugs(c *C) {
 
 // The auto-connect task will auto-connect slots with viable candidates.
 func (s *interfaceManagerSuite) TestDoSetupSnapSecurityAutoConnectsSlots(c *C) {
-	s.AssertsMock.MockModel(c, s.state, nil)
+	s.MockModel(c, s.state, nil)
 
 	// Mock the interface that will be used by the test
 	s.mockIfaces(c, &ifacetest.TestInterface{InterfaceName: "test"}, &ifacetest.TestInterface{InterfaceName: "test2"})
@@ -1636,7 +1748,7 @@ func (s *interfaceManagerSuite) TestDoSetupSnapSecurityAutoConnectsSlots(c *C) {
 
 // The auto-connect task will auto-connect slots with viable multiple candidates.
 func (s *interfaceManagerSuite) TestDoSetupSnapSecurityAutoConnectsSlotsMultiplePlugs(c *C) {
-	s.AssertsMock.MockModel(c, s.state, nil)
+	s.MockModel(c, s.state, nil)
 
 	// Mock the interface that will be used by the test
 	s.mockIfaces(c, &ifacetest.TestInterface{InterfaceName: "test"}, &ifacetest.TestInterface{InterfaceName: "test2"})
@@ -1739,7 +1851,7 @@ func (s *interfaceManagerSuite) TestDoSetupSnapSecurityNoAutoConnectSlotsIfAlter
 
 // The auto-connect task will auto-connect plugs with viable candidates also condidering snap declarations.
 func (s *interfaceManagerSuite) TestDoSetupSnapSecurityAutoConnectsDeclBased(c *C) {
-	s.AssertsMock.MockModel(c, s.state, nil)
+	s.MockModel(c, s.state, nil)
 
 	s.testDoSetupSnapSecurityAutoConnectsDeclBased(c, true, func(conns map[string]interface{}, repoConns []*interfaces.ConnRef) {
 		// Ensure that "test" plug is now saved in the state as auto-connected.
@@ -1776,7 +1888,7 @@ slots:
 	defer restore()
 	// Add the producer snap
 	s.mockIfaces(c, &ifacetest.TestInterface{InterfaceName: "test"}, &ifacetest.TestInterface{InterfaceName: "test2"})
-	s.AssertsMock.MockSnapDecl(c, "producer", "one-publisher", nil)
+	s.MockSnapDecl(c, "producer", "one-publisher", nil)
 	s.mockSnap(c, producerYaml)
 
 	// Initialize the manager. This registers the producer snap.
@@ -1784,7 +1896,7 @@ slots:
 
 	// Add a sample snap with a plug with the "test" interface which should be auto-connected.
 	if withDecl {
-		s.AssertsMock.MockSnapDecl(c, "consumer", "one-publisher", nil)
+		s.MockSnapDecl(c, "consumer", "one-publisher", nil)
 	}
 	snapInfo := s.mockSnap(c, consumerYaml)
 
@@ -1819,7 +1931,7 @@ slots:
 // in the model assertion fails an on-store constraint.
 func (s *interfaceManagerSuite) TestDoSetupSnapSecurityAutoConnectsDeclBasedDeviceScopeNoStore(c *C) {
 
-	s.AssertsMock.MockModel(c, s.state, nil)
+	s.MockModel(c, s.state, nil)
 
 	s.testDoSetupSnapSecurityAutoConnectsDeclBasedDeviceScope(c, func(conns map[string]interface{}, repoConns []*interfaces.ConnRef) {
 		// Ensure nothing is connected.
@@ -1833,7 +1945,7 @@ func (s *interfaceManagerSuite) TestDoSetupSnapSecurityAutoConnectsDeclBasedDevi
 // store in the model assertion fails an on-store constraint.
 func (s *interfaceManagerSuite) TestDoSetupSnapSecurityAutoConnectsDeclBasedDeviceScopeWrongStore(c *C) {
 
-	s.AssertsMock.MockModel(c, s.state, map[string]interface{}{
+	s.MockModel(c, s.state, map[string]interface{}{
 		"store": "other-store",
 	})
 
@@ -1849,7 +1961,7 @@ func (s *interfaceManagerSuite) TestDoSetupSnapSecurityAutoConnectsDeclBasedDevi
 // store in the model assertion passes an on-store constraint.
 func (s *interfaceManagerSuite) TestDoSetupSnapSecurityAutoConnectsDeclBasedDeviceScopeRightStore(c *C) {
 
-	s.AssertsMock.MockModel(c, s.state, map[string]interface{}{
+	s.MockModel(c, s.state, map[string]interface{}{
 		"store": "my-store",
 	})
 
@@ -1871,11 +1983,11 @@ func (s *interfaceManagerSuite) TestDoSetupSnapSecurityAutoConnectsDeclBasedDevi
 // on-store constraint.
 func (s *interfaceManagerSuite) TestDoSetupSnapSecurityAutoConnectsDeclBasedDeviceScopeWrongFriendlyStore(c *C) {
 
-	s.AssertsMock.MockModel(c, s.state, map[string]interface{}{
+	s.MockModel(c, s.state, map[string]interface{}{
 		"store": "my-substore",
 	})
 
-	s.AssertsMock.MockStore(c, s.state, "my-substore", map[string]interface{}{
+	s.MockStore(c, s.state, "my-substore", map[string]interface{}{
 		"friendly-stores": []interface{}{"other-store"},
 	})
 
@@ -1892,11 +2004,11 @@ func (s *interfaceManagerSuite) TestDoSetupSnapSecurityAutoConnectsDeclBasedDevi
 // on-store constraint.
 func (s *interfaceManagerSuite) TestDoSetupSnapSecurityAutoConnectsDeclBasedDeviceScopeFriendlyStore(c *C) {
 
-	s.AssertsMock.MockModel(c, s.state, map[string]interface{}{
+	s.MockModel(c, s.state, map[string]interface{}{
 		"store": "my-substore",
 	})
 
-	s.AssertsMock.MockStore(c, s.state, "my-substore", map[string]interface{}{
+	s.MockStore(c, s.state, "my-substore", map[string]interface{}{
 		"friendly-stores": []interface{}{"my-store"},
 	})
 
@@ -1924,13 +2036,13 @@ slots:
 	defer restore()
 	// Add the producer snap
 	s.mockIfaces(c, &ifacetest.TestInterface{InterfaceName: "test"})
-	s.AssertsMock.MockSnapDecl(c, "producer", "one-publisher", nil)
+	s.MockSnapDecl(c, "producer", "one-publisher", nil)
 	s.mockSnap(c, producerYaml)
 
 	// Initialize the manager. This registers the producer snap.
 	mgr := s.manager(c)
 
-	s.AssertsMock.MockSnapDecl(c, "consumer", "one-publisher", map[string]interface{}{
+	s.MockSnapDecl(c, "consumer", "one-publisher", map[string]interface{}{
 		"format": "3",
 		"plugs": map[string]interface{}{
 			"test": map[string]interface{}{
@@ -1971,7 +2083,7 @@ slots:
 // The setup-profiles task will only touch connection state for the task it
 // operates on or auto-connects to and will leave other state intact.
 func (s *interfaceManagerSuite) TestDoSetupSnapSecurityKeepsExistingConnectionState(c *C) {
-	s.AssertsMock.MockModel(c, s.state, nil)
+	s.MockModel(c, s.state, nil)
 
 	// Add an OS snap in place.
 	s.mockSnap(c, ubuntuCoreSnapYaml)
@@ -2245,7 +2357,7 @@ func (s *interfaceManagerSuite) TestSetupProfilesUsesFreshSnapInfo(c *C) {
 
 // auto-connect needs to setup security for connected slots after autoconnection
 func (s *interfaceManagerSuite) TestAutoConnectSetupSecurityForConnectedSlots(c *C) {
-	s.AssertsMock.MockModel(c, s.state, nil)
+	s.MockModel(c, s.state, nil)
 
 	// Add an OS snap.
 	coreSnapInfo := s.mockSnap(c, ubuntuCoreSnapYaml)
@@ -2435,7 +2547,7 @@ func (s *interfaceManagerSuite) TestDoRemove(c *C) {
 }
 
 func (s *interfaceManagerSuite) TestConnectTracksConnectionsInState(c *C) {
-	s.AssertsMock.MockModel(c, s.state, nil)
+	s.MockModel(c, s.state, nil)
 
 	s.mockIfaces(c, &ifacetest.TestInterface{InterfaceName: "test"}, &ifacetest.TestInterface{InterfaceName: "test2"})
 	s.mockSnap(c, consumerYaml)
@@ -2479,7 +2591,7 @@ func (s *interfaceManagerSuite) TestConnectTracksConnectionsInState(c *C) {
 }
 
 func (s *interfaceManagerSuite) TestConnectSetsUpSecurity(c *C) {
-	s.AssertsMock.MockModel(c, s.state, nil)
+	s.MockModel(c, s.state, nil)
 
 	s.mockIfaces(c, &ifacetest.TestInterface{InterfaceName: "test"}, &ifacetest.TestInterface{InterfaceName: "test2"})
 
@@ -2768,7 +2880,7 @@ func (s *interfaceManagerSuite) TestSetupProfilesDevModeMultiple(c *C) {
 }
 
 func (s *interfaceManagerSuite) TestCheckInterfacesDeny(c *C) {
-	s.AssertsMock.MockModel(c, s.state, nil)
+	s.MockModel(c, s.state, nil)
 
 	restore := assertstest.MockBuiltinBaseDeclaration([]byte(`
 type: base-declaration
@@ -2781,7 +2893,7 @@ slots:
 	defer restore()
 	s.mockIface(c, &ifacetest.TestInterface{InterfaceName: "test"})
 
-	s.AssertsMock.MockSnapDecl(c, "producer", "producer-publisher", nil)
+	s.MockSnapDecl(c, "producer", "producer-publisher", nil)
 	snapInfo := s.mockSnap(c, producerYaml)
 
 	s.state.Lock()
@@ -2810,7 +2922,7 @@ slots:
 }
 
 func (s *interfaceManagerSuite) TestCheckInterfacesAllow(c *C) {
-	s.AssertsMock.MockModel(c, s.state, nil)
+	s.MockModel(c, s.state, nil)
 
 	restore := assertstest.MockBuiltinBaseDeclaration([]byte(`
 type: base-declaration
@@ -2823,7 +2935,7 @@ slots:
 	defer restore()
 	s.mockIface(c, &ifacetest.TestInterface{InterfaceName: "test"})
 
-	s.AssertsMock.MockSnapDecl(c, "producer", "producer-publisher", map[string]interface{}{
+	s.MockSnapDecl(c, "producer", "producer-publisher", map[string]interface{}{
 		"format": "1",
 		"slots": map[string]interface{}{
 			"test": "true",
@@ -2837,7 +2949,7 @@ slots:
 }
 
 func (s *interfaceManagerSuite) TestCheckInterfacesDeviceScopeRightStore(c *C) {
-	s.AssertsMock.MockModel(c, s.state, map[string]interface{}{
+	s.MockModel(c, s.state, map[string]interface{}{
 		"store": "my-store",
 	})
 
@@ -2852,7 +2964,7 @@ slots:
 	defer restore()
 	s.mockIface(c, &ifacetest.TestInterface{InterfaceName: "test"})
 
-	s.AssertsMock.MockSnapDecl(c, "producer", "producer-publisher", map[string]interface{}{
+	s.MockSnapDecl(c, "producer", "producer-publisher", map[string]interface{}{
 		"format": "3",
 		"slots": map[string]interface{}{
 			"test": map[string]interface{}{
@@ -2870,7 +2982,7 @@ slots:
 }
 
 func (s *interfaceManagerSuite) TestCheckInterfacesDeviceScopeNoStore(c *C) {
-	s.AssertsMock.MockModel(c, s.state, nil)
+	s.MockModel(c, s.state, nil)
 
 	restore := assertstest.MockBuiltinBaseDeclaration([]byte(`
 type: base-declaration
@@ -2883,7 +2995,7 @@ slots:
 	defer restore()
 	s.mockIface(c, &ifacetest.TestInterface{InterfaceName: "test"})
 
-	s.AssertsMock.MockSnapDecl(c, "producer", "producer-publisher", map[string]interface{}{
+	s.MockSnapDecl(c, "producer", "producer-publisher", map[string]interface{}{
 		"format": "3",
 		"slots": map[string]interface{}{
 			"test": map[string]interface{}{
@@ -2901,7 +3013,7 @@ slots:
 }
 
 func (s *interfaceManagerSuite) TestCheckInterfacesDeviceScopeWrongStore(c *C) {
-	s.AssertsMock.MockModel(c, s.state, map[string]interface{}{
+	s.MockModel(c, s.state, map[string]interface{}{
 		"store": "other-store",
 	})
 
@@ -2916,7 +3028,7 @@ slots:
 	defer restore()
 	s.mockIface(c, &ifacetest.TestInterface{InterfaceName: "test"})
 
-	s.AssertsMock.MockSnapDecl(c, "producer", "producer-publisher", map[string]interface{}{
+	s.MockSnapDecl(c, "producer", "producer-publisher", map[string]interface{}{
 		"format": "3",
 		"slots": map[string]interface{}{
 			"test": map[string]interface{}{
@@ -2934,11 +3046,11 @@ slots:
 }
 
 func (s *interfaceManagerSuite) TestCheckInterfacesDeviceScopeRightFriendlyStore(c *C) {
-	s.AssertsMock.MockModel(c, s.state, map[string]interface{}{
+	s.MockModel(c, s.state, map[string]interface{}{
 		"store": "my-substore",
 	})
 
-	s.AssertsMock.MockStore(c, s.state, "my-substore", map[string]interface{}{
+	s.MockStore(c, s.state, "my-substore", map[string]interface{}{
 		"friendly-stores": []interface{}{"my-store"},
 	})
 
@@ -2953,7 +3065,7 @@ slots:
 	defer restore()
 	s.mockIface(c, &ifacetest.TestInterface{InterfaceName: "test"})
 
-	s.AssertsMock.MockSnapDecl(c, "producer", "producer-publisher", map[string]interface{}{
+	s.MockSnapDecl(c, "producer", "producer-publisher", map[string]interface{}{
 		"format": "3",
 		"slots": map[string]interface{}{
 			"test": map[string]interface{}{
@@ -2971,11 +3083,11 @@ slots:
 }
 
 func (s *interfaceManagerSuite) TestCheckInterfacesDeviceScopeWrongFriendlyStore(c *C) {
-	s.AssertsMock.MockModel(c, s.state, map[string]interface{}{
+	s.MockModel(c, s.state, map[string]interface{}{
 		"store": "my-substore",
 	})
 
-	s.AssertsMock.MockStore(c, s.state, "my-substore", map[string]interface{}{
+	s.MockStore(c, s.state, "my-substore", map[string]interface{}{
 		"friendly-stores": []interface{}{"other-store"},
 	})
 
@@ -2990,7 +3102,7 @@ slots:
 	defer restore()
 	s.mockIface(c, &ifacetest.TestInterface{InterfaceName: "test"})
 
-	s.AssertsMock.MockSnapDecl(c, "producer", "producer-publisher", map[string]interface{}{
+	s.MockSnapDecl(c, "producer", "producer-publisher", map[string]interface{}{
 		"format": "3",
 		"slots": map[string]interface{}{
 			"test": map[string]interface{}{
@@ -3247,7 +3359,7 @@ plugs:
 }
 
 func (s *interfaceManagerSuite) TestAutoConnectDuringCoreTransition(c *C) {
-	s.AssertsMock.MockModel(c, s.state, nil)
+	s.MockModel(c, s.state, nil)
 
 	// Add both the old and new core snaps
 	s.mockSnap(c, ubuntuCoreSnapYaml)
@@ -3334,7 +3446,7 @@ func makeAutoConnectChange(st *state.State, plugSnap, plug, slotSnap, slot strin
 }
 
 func (s *interfaceManagerSuite) TestUndoConnect(c *C) {
-	s.AssertsMock.MockModel(c, s.state, nil)
+	s.MockModel(c, s.state, nil)
 
 	s.mockIfaces(c, &ifacetest.TestInterface{InterfaceName: "test"})
 	s.manager(c)
@@ -3504,7 +3616,7 @@ func (s *interfaceManagerSuite) TestConnectErrorMissingSlotOnAutoConnect(c *C) {
 }
 
 func (s *interfaceManagerSuite) TestConnectHandlesAutoconnect(c *C) {
-	s.AssertsMock.MockModel(c, s.state, nil)
+	s.MockModel(c, s.state, nil)
 
 	s.mockIfaces(c, &ifacetest.TestInterface{InterfaceName: "test"})
 	_ = s.manager(c)
@@ -3574,7 +3686,7 @@ func (s *interfaceManagerSuite) TestRegenerateAllSecurityProfilesWritesSystemKey
 }
 
 func (s *interfaceManagerSuite) TestAutoconnectSelf(c *C) {
-	s.AssertsMock.MockModel(c, s.state, nil)
+	s.MockModel(c, s.state, nil)
 
 	s.mockIfaces(c, &ifacetest.TestInterface{InterfaceName: "test"})
 	s.mockSnap(c, selfconnectSnapYaml)
@@ -3999,9 +4111,9 @@ func (s *interfaceManagerSuite) TestDisconnectInterfacesRetrySetupProfiles(c *C)
 
 func (s *interfaceManagerSuite) setupGadgetConnect(c *C) {
 	s.mockIfaces(c, &ifacetest.TestInterface{InterfaceName: "test"})
-	s.AssertsMock.MockSnapDecl(c, "consumer", "publisher1", nil)
+	s.MockSnapDecl(c, "consumer", "publisher1", nil)
 	s.mockSnap(c, consumerYaml)
-	s.AssertsMock.MockSnapDecl(c, "producer", "publisher2", nil)
+	s.MockSnapDecl(c, "producer", "publisher2", nil)
 	s.mockSnap(c, producerYaml)
 
 	gadgetInfo := s.mockSnap(c, `name: gadget
@@ -4151,9 +4263,9 @@ func (s *interfaceManagerSuite) TestGadgetConnectSkipUnknown(c *C) {
 	defer r1()
 
 	s.mockIfaces(c, &ifacetest.TestInterface{InterfaceName: "test"})
-	s.AssertsMock.MockSnapDecl(c, "consumer", "publisher1", nil)
+	s.MockSnapDecl(c, "consumer", "publisher1", nil)
 	s.mockSnap(c, consumerYaml)
-	s.AssertsMock.MockSnapDecl(c, "producer", "publisher2", nil)
+	s.MockSnapDecl(c, "producer", "publisher2", nil)
 	s.mockSnap(c, producerYaml)
 
 	s.manager(c)
@@ -4206,11 +4318,11 @@ func (s *interfaceManagerSuite) TestGadgetConnectHappyPolicyChecks(c *C) {
 	r1 := release.MockOnClassic(false)
 	defer r1()
 
-	s.AssertsMock.MockModel(c, s.state, nil)
+	s.MockModel(c, s.state, nil)
 
 	s.mockSnap(c, coreSnapYaml)
 
-	s.AssertsMock.MockSnapDecl(c, "foo", "publisher1", nil)
+	s.MockSnapDecl(c, "foo", "publisher1", nil)
 	s.mockSnap(c, `name: foo
 version: 1.0
 plugs:
